script midpoint
